# [POC][wip] faster `DefaultEngine` parquet reads