--- a/Cargo.toml
+++ b/Cargo.toml
@@ -22,17 +22,17 @@ rust-version = "1.80"
 version = "0.5.0"
 
 [workspace.dependencies]
-arrow = { version = ">=53, <54" }
-arrow-arith = { version = ">=53, <54" }
-arrow-array = { version = ">=53, <54" }
-arrow-buffer = { version = ">=53, <54" }
-arrow-cast = { version = ">=53, <54" }
-arrow-data = { version = ">=53, <54" }
-arrow-ord = { version = ">=53, <54" }
-arrow-json = { version = ">=53, <54" }
-arrow-select = { version = ">=53, <54" }
-arrow-schema = { version = ">=53, <54" }
-parquet = { version = ">=53, <54", features = ["object_store"] }
+arrow = { version = ">=53.3, <54" }
+arrow-arith = { version = ">=53.3, <54" }
+arrow-array = { version = ">=53.3, <54" }
+arrow-buffer = { version = ">=53.3, <54" }
+arrow-cast = { version = ">=53.3, <54" }
+arrow-data = { version = ">=53.3, <54" }
+arrow-ord = { version = ">=53.3, <54" }
+arrow-json = { version = ">=53.3, <54" }
+arrow-select = { version = ">=53.3, <54" }
+arrow-schema = { version = ">=53.3, <54" }
+parquet = { version = ">=53.3, <54", features = ["object_store"] }
 object_store = { version = ">=0.11, <0.12" }
 hdfs-native-object-store = "0.12.0"
 hdfs-native = "0.10.0"