## TLDR
This PR is a POC/exploration on speeding up `DefaultEngine` parquet reads. The current implementation is unfortunately relatively complicated and despite a solid amount of async code appears to be serially reading all parquet files. There are two main outcomes from this exploration:
1. Empirically showed that the `DefaultParquetHandler::read_parquet_files` implementation is indeed serial but can be trivially made async/concurrent using typical async code and tokio tasks. In this case we can fire off all IO requests (up to some limit) and then bridge the async-to-sync boundary with an mpsc channel.
2. Within our `Scan::execute` implementation we only ever pass a singular `FileMeta` to `read_parquet_files` thereby immediately limiting any concurrency implemented in the engine. This wasn't explored further in this PR but likely takes some more design work since there is a requirement to colocate a parquet file's partition values with the outcome of the parquet read. This doesn't seem to fit nicely in the existing API.

## Details
### Need for a better `DefaultParquetHandler::read_parquet_files`
The POC here gives an alternative (strawman) as `AsyncParquetHandler` which simply launches a tokio task for each parquet file to read. The rudimentary tests (that only work on my machine lol) show that the existing implementation serially reads each file (despite `readahead = 10`) and the new implementation indeed fires of all IO immediately. (tests simulated high IO latency via 'sleep')

#### Future work
- understand why the existing implementation is serial (obviously a bug and not intended)
- consider productionization of something like the `AsyncParquetHandler`
- include benchmarks/other substantiation of changes made in this area
- consider if/when/how to integrate with various async runtimes. if the consumer of the kernel is also an async rust user, it may be beneficial to propagate a runtime handle so that we don't end up with competing runtimes.
- consider overall execution design in the `DefaultEngine`: do we want multiple runtimes? (IO runtime and CPU-bound runtime?)

### Need for a better `Scan::execute` implementation
Currently, within `Scan::execute` we take the `ScanFile` iterator and sequentially call `read_parquet_files()` on each one:
```rust
    let result = scan_files_iter
        .map(move |scan_file| -> DeltaResult<_> {
            // ... [snip] ...
            let read_result_iter = engine.get_parquet_handler().read_parquet_files(
                &[meta],
                global_state.physical_schema.clone(),
                physical_predicate.clone(),
            )?;
            // ... [snip] ...
            Ok(read_result_iter.map(move |read_result| -> DeltaResult<_> {
                // to transform the physical data into the correct logical form
                let logical = transform_to_logical_internal(...);
                // ... [snip] ...
                Ok(result)
            }))
        })
        .flatten_ok()
        .map(|x| x?);
```
Instead, ideally we would pass many parquet files required for the scan at once and then let the engine decide how to schedule IO. This is unfortunately not easy to coordinate since we have per-file state that must propagate 'through' the read_parquet_file API. That is, if we colocate all parquet files in the scan (spanning multiple partitions) we must somehow align all the data we read with the corresponding partition values. I didn't delve too much further here but flagging this for optimization soon.

#### Future work
- consider all the spots we do parquet file reads, do we pass in multiple files to leverage concurrency implemented by the engine?
- how can we fix the execute implementation to allow for reading all parquet files at once and propagating necessary per-file information?