@scovich 2025-01-06
> @zachschuermann You mention in the PR description:
> 
> > Instead, ideally we would pass many parquet files required for the scan at once and then let the engine decide how to schedule IO. This is unfortunately not easy to coordinate since we have per-file state that must propagate 'through' the read_parquet_file API. That is, if we colocate all parquet files in the scan (spanning multiple partitions) we must somehow align all the data we read with the corresponding partition values. I didn't delve too much further here but flagging this for optimization soon.
> 
> I'm not sure I fully get this. ScanFile already holds the partition value for that file, so even if there are multiple parquet files that belong to a partition, each of them will have their own ScanFile telling them what their partition value is. Given that, it should be pretty easy to parallelize: (GlobalScanState + ScanFile) is everything you need.

This was also my understanding of how scans worked -- or at least how the original design wanted them to work?